pyproject.toml
[project]
name = "polymarket-agents"
version = "0.1.0"
description = "Realtime, low-latency Calfkit agents trading on Polymarket"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "calfkit>=0.2.3",
    "pyyaml>=6.0.3",
    "websockets>=13.0",
    "pydantic-settings>=2.0",
]
